Move a few test helpers to more common place
The users.tests.conftest module seems to be most used place for common
test helpers.
