+++ tunnistamo/tests/conftest.py
@@ -1,24 +1,19 @@
 import json
-import time
 from urllib.parse import parse_qs, urlparse
 
 import jwt
 import pytest
-from Cryptodome.PublicKey.RSA import importKey
 from django.test import Client as TestClient
 from django.urls import reverse
 from django.utils.crypto import get_random_string
-from jwkest.jwk import RSAKey as jwk_RSAKey
-from jwkest.jws import JWS
 from oidc_provider.models import RESPONSE_TYPE_CHOICES
 from oidc_provider.models import Client as OidcClient
-from oidc_provider.models import ResponseType, RSAKey
+from oidc_provider.models import ResponseType
 
-from auth_backends.helsinki_tunnistus_suomifi import HelsinkiTunnistus
 from oidc_apis.models import Api, ApiDomain, ApiScope
 from oidc_apis.views import get_api_tokens_view
 from users.tests.conftest import (  # noqa
-    DummyOidcBackendBase, oidcclient_factory, tunnistamosession_factory, user, usersocialauth_factory
+    DummyFixedOidcBackend, oidcclient_factory, tunnistamosession_factory, user, usersocialauth_factory
 )
 
 
@@ -49,63 +44,6 @@ def rsa_key():
     return create_rsa_key()
 
 
-def create_id_token(backend, **kwargs):
-    kwargs.setdefault('iss', backend.oidc_config().get('issuer'))
-    kwargs.setdefault('sub', get_random_string())
-    kwargs.setdefault('aud', backend.setting('KEY'))
-    kwargs.setdefault('azp', backend.setting('KEY'))
-    kwargs.setdefault('exp', int(time.time()) + 60 * 5)
-    kwargs.setdefault('iat', int(time.time()) - 10)
-    kwargs.setdefault('jti', get_random_string())
-    kwargs.setdefault('name', get_random_string())
-    kwargs.setdefault('given_name', get_random_string())
-    kwargs.setdefault('family_name', get_random_string())
-
-    keys = []
-    for rsakey in RSAKey.objects.all():
-        keys.append(jwk_RSAKey(key=importKey(rsakey.key), kid=rsakey.kid))
-
-    _jws = JWS(kwargs, alg='RS256')
-    return _jws.sign_compact(keys)
-
-
-class DummyFixedOidcBackend(DummyOidcBackendBase):
-    """Dummy OIDC social auth backend that returns fixed access and id tokens"""
-    name = 'dummyfixedoidcbackend'
-
-    def user_data(self, access_token, *args, **kwargs):
-        return {
-            'email_verified': False,
-            'family_name': 'User',
-            'given_name': 'Test',
-            'name': 'Test User',
-            'sub': '00000000-0000-4000-b000-000000000000'
-        }
-
-    def get_json(self, url, *args, **kwargs):
-        if url == self.oidc_config()['token_endpoint']:
-            nonce = self.get_and_store_nonce(self.authorization_url(), get_random_string())
-            id_token = create_id_token(
-                self,
-                nonce=nonce,
-                sub='00000000-0000-4000-b000-000000000000',
-                email_verified=False,
-                name='Test User',
-                given_name='User',
-                family_name='Test',
-                loa='substantial',
-            )
-
-            return {
-                'access_token': 'access_token_abcd123',
-                'id_token': id_token,
-            }
-
-        return super().get_json(url, *args, **kwargs)
-
-    get_loa = HelsinkiTunnistus.get_loa
-
-
 def social_login(settings, test_client=None, trust_loa=True):
     """Authenticate a user using a dummy social auth backend
 
@@ -117,7 +55,7 @@ def social_login(settings, test_client=None, trust_loa=True):
     create_rsa_key()
 
     settings.AUTHENTICATION_BACKENDS = settings.AUTHENTICATION_BACKENDS + (
-        'tunnistamo.tests.conftest.DummyFixedOidcBackend',
+        'users.tests.conftest.DummyFixedOidcBackend',
     )
 
     settings.SOCIAL_AUTH_DUMMYFIXEDOIDCBACKEND_KEY = 'tunnistamo'
